+++ lib/interface-types/src/errors.rs
@@ -8,6 +8,7 @@ use crate::{
 use std::{
     error::Error,
     fmt::{self, Display, Formatter},
+    num::TryFromIntError,
     result::Result,
     string::{self, ToString},
 };
@@ -153,6 +154,12 @@ pub enum InstructionErrorKind {
     /// The string contains invalid UTF-8 encoding.
     String(string::FromUtf8Error),
 
+    /// Out of range integral type conversion attempted.
+    NegativeValue {
+        /// The variable name that triggered the error.
+        subject: &'static str,
+    },
+
     /// The type doesn't exist.
     TypeIsMissing {
         /// The type index.
@@ -237,6 +244,12 @@ impl Display for InstructionErrorKind {
 
             Self::String(error) => write!(formatter, "{}", error),
 
+            Self::NegativeValue { subject } => write!(
+                formatter,
+                "attempted to convert `{}` but it appears to be a negative value",
+                subject
+            ),
+
             Self::TypeIsMissing { type_index } => write!(
                 formatter,
                 "the type `{}` doesn't exist",
@@ -249,5 +262,11 @@ impl Display for InstructionErrorKind {
                 received_kind, expected_kind
             ),
         }
+    }
+}
+
+impl From<(TryFromIntError, &'static str)> for InstructionErrorKind {
+    fn from((_, subject): (TryFromIntError, &'static str)) -> Self {
+        InstructionErrorKind::NegativeValue { subject }
     }
 }